requirements.txt
llama-index-llms-openai
python-dotenv==1.0.1
streamlit==1.36.0
llama_index==0.11.19
importlib==1.0.4